Fix "global options already processed" error
AM_INIT_AUTOMAKE() cannot be called twice.  Merge those two to one.

Also, Makefile.in should not be in the git repo.  The file is auto
generated by autotools from Makefile.am, which is the one to be
in the repo.  I'd go for Non-recursive Automake but that's another
story.